+++ pkg/fs/local_file_system_windows.go
@@ -63,3 +63,7 @@ func (fs *localFileSystem) SyncPath(_ string) {}
 func syncFile(_ *os.File) error {
 	return nil
 }
+
+func CompareINode(srcPath, destPath string) error {
+	return nil
+}